Refs #3286. Big refactoring of filter slider code.
- Created a new package.
- RangeChooser<N> is now a generic type, so values of different types
can't accidentally be mixed.

+++ filter2-impl/src/main/java/org/cytoscape/filter/internal/filters/column/ColumnFilterController.java
@@ -3,6 +3,6 @@ package org.cytoscape.filter.internal.filters.column;
 public interface ColumnFilterController {
 	ColumnFilter getFilter();
 	void columnsChanged(ColumnFilterView view);
-	void setSliderBounds(long min, long max);
+	void setSliderBounds(int min, int max);
 	void setSliderBounds(double min, double max);
 }
